design/requirements.txt
fastapi
uvicorn
langchain-ollama
langchain-community
pydantic
sqlmodel
